# docs/openapi.yaml
openapi: 3.1.0
info:
  title: سیستم مدیریت پرسنل خط ۱ مترو تهران
  description: |
    REST API برای سیستم مدیریت پرسنل خط ۱ مترو تهران.
    شامل احراز هویت، دفتر تلفن، مدیریت شیفت، بخشنامه ایمنی، تیکتینگ و مدیریت کاربران.
  version: 1.0.0
  contact:
    name: تیم توسعه مترو خط ۱
servers:
  - url: /api
    description: سرور توسعه
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        nationalId:
          type: string
        name:
          type: string
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, pending, suspended]
        customFields:
          type: object
          nullable: true
        role:
          type: object
          properties:
            id:
              type: string
            key:
              type: string
            name:
              type: string
        createdAt:
          type: string
          format: date-time
...
paths:
...
  /auth/login:
    post:
      tags: [احراز هویت]
      summary: ورود کاربر
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nationalId, password]
              properties:
                nationalId:
                  type: string
                  example: "0000000000"
                password:
                  type: string
                  example: "admin123"
      responses:
        '200':
          description: موفق
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
                      accessToken:
                        type: string
                      refreshToken:
                        type: string
...
  /profile:
    get:
      tags: [پروفایل]
      summary: دریافت پروفایل کاربر جاری
      responses:
        '200':
          description: پروفایل کاربر
    patch:
      tags: [پروفایل]
      summary: بروزرسانی پروفایل
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                phone:
                  type: string
                email:
                  type: string
                vehicles:
                  type: array
                  items:
                    type: object
                    properties:
                      plateNum1:
                        type: string
                      plateLetter:
                        type: string
                      plateNum2:
                        type: string
                      plateCity:
                        type: string
                      carType:
                        type: string
                      carColor:
                        type: string
      responses:
        '200':
          description: بروزرسانی موفق

  # ── Shifts ──────────────────────────────────────────
...